app/requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0